InvoideDate refactoring.
